refactor(build): copy files from node_modules instead of Npm.depends
since the latter also bundles Ionic files into the Atmosphere package, with assets (so they are duplicated there).
Instead, simply load Ionic from npm into node_modules, copy to a local "dist" folder (not committed, but bundled into the package) and link them in "package.js". This way, we read from local File System, instead of relying on an external 3rd party service to list the files.
Removing the node_modules folder at the end of the process, as recommended in the Meteor package authoring guidelines.