fix(plugin): compatibility with Blaze
get rid of the HTML file so that we no longer need to specify a dependency on "static-html" or "blaze-html-templates" Atmosphere package.
When we specify such a dependency, it gets automatically installed in the consumer project.
Unfortunately the 2 are mutually exclusive, since they register as compilers for "HTML" files, and Meteor cannot accept more than 1 compiler for a given file type.
Therefore the plugin had to be specified either for "static-html" (i.e. for use with Vue, React or Angular) or "blaze-html-templates" (i.e. for Blaze).
By using the "data-resources-url" undocumented feature of @ionic/core / StencilJS, we can hard-code the path to assets instead of relying on ionic / stencil loader to guess it from the <script> that loads "ionic.js".
Therefore we are now immune to bundling, and we no longer need to import "ionic.js" the old way (i.e. through a <script>), that was the root cause for needing a HTML file and specifying a corresponding compiler.