feat(demoVue): working demo with Vue
difference compared to Blaze demo: the mobile preview supports in-app mode switching.
It actually posts a message to the top window (should be parent?) and delegates mode switching to it, so that the latter can adjust the containing figure display (notch for iOS), if there is such a window. Otherwise, reloads itself as usual.